+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIManagerConfiguration.java
@@ -488,7 +488,7 @@ public class APIManagerConfiguration {
                             " configuration under [apim.redis_config] section in deployment.toml");
                 }
                 if (minGatewayCount != null) {
-                    redisConfig.setMinGatewayCount(Integer.parseInt(minGatewayCount.getText()));
+                    redisConfig.setMinGatewayCount(Long.parseLong(minGatewayCount.getText()));
                 }
                 if (keyLockRetrievalTimeout != null) {
                     redisConfig.setKeyLockRetrievalTimeout(Integer.parseInt(keyLockRetrievalTimeout.getText()));
